Windows: avoid calling select() with more than FD_SETSIZE fds
This is similar to "PR#5563: harden Unix.select against file descriptors above FD_SETSIZE"
except that we can still handle these larger lists using the more generic
(but more complicated) codepath.

The MSDN documentation for select states that "The variable FD_SETSIZE
determines the maximum number of descriptors in a set.":

  https://msdn.microsoft.com/en-gb/library/windows/desktop/ms740141(v=vs.85).aspx

If we have a list of fds which is too large to be used in select(),
fall back to the generic codepath which is used for select()ing on multiple
types of handles at once.

Note the default FD_SETSIZE is 64, so it's fairly easy to hit this limit.

Signed-off-by: David Scott <dave@recoil.org>